refactor telescope extensions
move them all to plugins/telescope/
