+++ .config/nvim/lua/plugins/gitsigns.lua
@@ -1,16 +1,12 @@
 --=============================================================================
 -------------------------------------------------------------------------------
---                                                                GITSIGNS.NVIM
+--                                                                     GITSIGNS
 --=============================================================================
 -- https://github.com/lewis6991/gitsigns.nvim
---_____________________________________________________________________________
+-- https://github.com/samjswill/nvim-unception
+--[[___________________________________________________________________________
 
---[[
 Git decorators.
-
-Added lines's numbers are shown in green,
-removed in red and modified in blue.
-
 Show the current line's git blame with a 700ms delay.
 
 Keymaps:
@@ -20,21 +16,13 @@ Keymaps:
    - <leader>gu: Unstage current buffer
    - <leader>gr: Reset current buffer
 
-   - <leader>gg Git status
-      see :h telescope.builtin.git_status
-   - <leader>gb: Git branches
-      see :h telescope.builtin.git_branches
-   - <leader>gl Git log
-      see :h telescope.builtin.git_commits
-   - <leader>gS: Git stash
-          - <CR> to apply the selected stash
-   - <leader>gf: Git files
+Add custom git commands:
    - <leader>gc: Git commit
    - <leader>gP: Git push
    - <leader>gp: Git pull
    - <leader>g or (:Git) :  Custom git command
 
---]]
+------------------------------------------------------------------------------]]
 
 local M = {
   "lewis6991/gitsigns.nvim",
@@ -43,6 +31,19 @@ local M = {
     "samjwill/nvim-unception",
   },
 }
+
+function M.init()
+  vim.keymap.set("n", "<leader>g", M.git_command, {})
+  vim.api.nvim_create_user_command("Git", function()
+    M.git_command()
+  end, {})
+  vim.keymap.set("n", "<leader>gc", M.git_commit, {})
+  vim.keymap.set("n", "<leader>gP", M.git_push, {})
+  vim.keymap.set("n", "<leader>gp", M.git_pull, {})
+
+  vim.g.unception_block_while_host_edits = true
+  vim.g.unception_enable_flavor_text = false
+end
 
 function M.config()
   local gitsigns = require "gitsigns"
@@ -67,7 +68,6 @@ function M.config()
       vim.keymap.set("n", "<leader>gr", gs.reset_buffer, opts)
     end,
   }
-
   vim.api.nvim_create_autocmd("User", {
     pattern = "UnceptionEditRequestReceived",
     callback = function()
@@ -83,65 +83,6 @@ end
 
 local call_git_command
 
-function M.init()
-  vim.keymap.set("n", "<leader>g", M.git_command, {})
-  vim.api.nvim_create_user_command("Git", function()
-    M.git_command()
-  end, {})
-
-  vim.keymap.set("n", "<leader>gl", M.git_commits, {})
-  vim.keymap.set("n", "<leader>gb", M.git_branches, {})
-  vim.keymap.set("n", "<leader>gg", M.git_status, {})
-  vim.keymap.set("n", "<leader>gf", M.git_files, {})
-  vim.keymap.set("n", "<leader>gS", M.git_stash, {})
-  vim.keymap.set("n", "<leader>gc", M.git_commit, {})
-  vim.keymap.set("n", "<leader>gP", M.git_push, {})
-  vim.keymap.set("n", "<leader>gp", M.git_pull, {})
-
-  vim.g.unception_block_while_host_edits = true
-  vim.g.unception_enable_flavor_text = false
-end
-
-M.theme = "ivy"
-
-local wrap_opts
-
-function M.git_commits(opts)
-  local telescope = require "telescope.builtin"
-
-  opts = wrap_opts(opts)
-  telescope.git_commits(opts)
-end
-
-function M.git_branches(opts)
-  local telescope = require "telescope.builtin"
-
-  opts = wrap_opts(opts)
-  telescope.git_branches(opts)
-end
-
-function M.git_files(opts)
-  local telescope = require "telescope.builtin"
-
-  opts = wrap_opts(opts)
-  telescope.git_files(opts)
-end
-
-function M.git_stash(opts)
-  local telescope = require "telescope.builtin"
-
-  opts = wrap_opts(opts)
-  telescope.git_stash(opts)
-end
-
-local attach_git_status_mappings
-function M.git_status(opts)
-  local telescope = require "telescope.builtin"
-  opts = wrap_opts(opts)
-  opts.attach_mappings = attach_git_status_mappings
-  telescope.git_status(opts)
-end
-
 function M.git_command()
   call_git_command "git "
 end
@@ -156,22 +97,6 @@ end
 
 function M.git_commit()
   call_git_command("git commit ", "-a")
-end
-
-function attach_git_status_mappings(_, map)
-  local actions = require "telescope.actions"
-  actions.select_default:replace(actions.git_staging_toggle)
-  map("n", "e", actions.file_edit)
-  map("i", "<C-e>", actions.file_edit)
-  map("i", "<Tab>", actions.move_selection_next)
-  map("i", "<Tab>", actions.move_selection_next)
-  map("n", "<Tab>", actions.move_selection_next)
-  map("i", "<S-Tab>", actions.move_selection_previous)
-  map("n", "<S-Tab>", actions.move_selection_previous)
-  map("n", "s", actions.git_staging_toggle)
-  map("i", "<C-s>", actions.git_staging_toggle)
-  map("n", "<C-s>", actions.git_staging_toggle)
-  return true
 end
 
 local run_cmd
@@ -238,18 +163,6 @@ function run_cmd(cmd)
       })
     end
   end)
-end
-
-wrap_opts = function(opts)
-  opts = vim.tbl_extend(
-    "force",
-    require("telescope.themes")["get_" .. M.theme](),
-    opts or {}
-  )
-  if not opts.selection_strategy then
-    opts.selection_strategy = "row"
-  end
-  return opts
 end
 
 return M
